+++ src/rotty/block.rs
@@ -64,7 +64,7 @@ impl Block {
         }
     }
 
-    pub fn width(&self) -> u32 {
+    pub fn width(&self) -> u16 {
         match self {
             Block::Image(img) => img.width,
             Block::Join { dir, blocks } => match dir {
@@ -77,7 +77,7 @@ impl Block {
     }
 
     // Eat your heart out, npm.
-    pub fn left_pad(self, width: u32) -> Self {
+    pub fn left_pad(self, width: u16) -> Self {
         let block_width = self.width();
         if width <= block_width {
             self
@@ -110,12 +110,12 @@ pub struct Image {
     pub attrs: style::Attributes,
 
     pub text: String,
-    pub width: u32,
+    pub width: u16,
     pub align: TextAlign,
 }
 
 impl Image {
-    pub fn new(t: &str, width: u32, align: TextAlign) -> Self {
+    pub fn new(t: &str, width: u16, align: TextAlign) -> Self {
         Image {
             text: t.to_string(),
             width,
